Fix infinite recursion in ErrCode.Error method.
Fixes #9

+++ errors.go
@@ -46,7 +46,7 @@ func (e ErrCode) String() string {
 	if s, ok := errCodeName[e]; ok {
 		return s
 	}
-	return fmt.Sprintf("unknown error code 0x%x", e)
+	return fmt.Sprintf("unknown error code 0x%x", uint32(e))
 }
 
 // ConnectionError is an error that results in the termination of the
